feat: add Browser.get|setWindowBounds and Page.windowId methods

+++ src/main/java/com/ruiyun/jvppeteer/common/WindowBounds.java
@@ -1,41 +1,41 @@
 package com.ruiyun.jvppeteer.common;
 
 public class WindowBounds {
-    public Double left;
-    public Double top;
-    public Double width;
-    public Double height;
+    public Integer left;
+    public Integer top;
+    public Integer width;
+    public Integer height;
     public WindowState windowState;
 
-    public Double getLeft() {
+    public Integer getLeft() {
         return left;
     }
 
-    public void setLeft(Double left) {
+    public void setLeft(Integer left) {
         this.left = left;
     }
 
-    public Double getTop() {
+    public Integer getTop() {
         return top;
     }
 
-    public void setTop(Double top) {
+    public void setTop(Integer top) {
         this.top = top;
     }
 
-    public Double getWidth() {
+    public Integer getWidth() {
         return width;
     }
 
-    public void setWidth(Double width) {
+    public void setWidth(Integer width) {
         this.width = width;
     }
 
-    public Double getHeight() {
+    public Integer getHeight() {
         return height;
     }
 
-    public void setHeight(Double height) {
+    public void setHeight(Integer height) {
         this.height = height;
     }
